+++ CHANGELOG.md
@@ -1,5 +1,57 @@
 # Changelog
 
+## v0.1.24
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.23...v0.1.24)
+
+## v0.1.23
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.22...v0.1.23)
+
+### 🚀 Enhancements
+
+- Add icon slashCircle ([e7e6eaf](https://github.com/NPO-Automotive/sep_yui/commit/e7e6eaf))
+
+## v0.1.22
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.21...v0.1.22)
+
+### 🚀 Enhancements
+
+- Color contrast improved of table row ([d668f6c](https://github.com/NPO-Automotive/sep_yui/commit/d668f6c))
+
+### 🩹 Fixes
+
+- Return background color to back ([71fdfee](https://github.com/NPO-Automotive/sep_yui/commit/71fdfee))
+
+## v0.1.21
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.20...v0.1.21)
+
+### 🚀 Enhancements
+
+- Add root color ([d9385a3](https://github.com/NPO-Automotive/sep_yui/commit/d9385a3))
+
+## v0.1.20
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.19...v0.1.20)
+
+### 🚀 Enhancements
+
+- Fix backgroun in table row and fix get gup in tooltip ([b83de5d](https://github.com/NPO-Automotive/sep_yui/commit/b83de5d))
+
+### 🩹 Fixes
+
+- Fix style position hint ([9dc0ee1](https://github.com/NPO-Automotive/sep_yui/commit/9dc0ee1))
+
+## v0.1.19
+
+[compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.18...v0.1.19)
+
+### 🚀 Enhancements
+
+- Added integer type number in inputNumber ([b680a8c](https://github.com/NPO-Automotive/sep_yui/commit/b680a8c))
+
 ## v0.1.18
 
 [compare changes](https://github.com/NPO-Automotive/sep_yui/compare/v0.1.17...v0.1.18)